KIN-36 Checks a local commit hash against the latest commit for the geth framework repo and updates if needed.

+++ scripts/get_geth.sh
@@ -1,15 +1,24 @@
 #!/bin/bash
 
-# downloads and installs a pre-built Geth framework with an i386 slice
+# updates / downloads and installs a pre-built Geth framework with an i386 slice
 
-echo "Downloading Geth..."
-target_dir=$PWD/KinSDK/Geth.framework
-rm -r KinSDK/Geth.framework 2>/dev/null
-cd /tmp
-rm -r Geth-iOS 2>/dev/null
-git clone --depth=1 https://github.com/kinfoundation/Geth-iOS.git
-cd Geth-iOS
-unzip -a Geth.framework.zip
-cp -r Geth.framework $target_dir
-cd ..
-rm -rf Geth-iOS 2>/dev/null
+export LOCAL_GETH_COMMIT=$(touch geth-commit && cat geth-commit )
+export REMOTE_GETH_COMMIT=$(git ls-remote https://github.com/kinfoundation/Geth-iOS.git | grep HEAD | awk '{ print $1}')
+
+curr_dir=$PWD
+target_dir=$curr_dir/KinSDK/Geth.framework
+
+if [ $LOCAL_GETH_COMMIT != $REMOTE_GETH_COMMIT ] || [ ! -r $target_dir ] ; then
+  echo "Geth library out of date or missing. Updating..."
+  rm -r KinSDK/Geth.framework 2>/dev/null
+  cd /tmp
+  rm -r -f Geth-iOS
+  git clone --depth=1 https://github.com/kinfoundation/Geth-iOS.git
+  cd Geth-iOS
+  unzip -a Geth.framework.zip
+  cp -r Geth.framework $target_dir
+  cd ..
+  rm -rf Geth-iOS 2>/dev/null
+  cd $curr_dir
+  echo $REMOTE_GETH_COMMIT > geth-commit
+fi
